Created Enemy Spawner
The enemy spawner takes in a min and max amount of enemies per spawn, a min and max amount of time to wait for spawn, a spawn offset so they don't spawn on each other if the user wants, and the enemy it wants to spawn. I also added the take damage function to player controller.