player detection sphere corrections
Enemies used to have trigger colliders for finding the player attached to the same object as the hitbox, meaning the player could shoot that trigger to damage the enemy. The trigger spheres have been moved to their own objects and scripts have been modified to allow that trigger to call the OnTrigger functions of the EnemyAI, EnemyAIWithRoam, and RobotEnemyAI scripts.

Additionally, several triggers in the library scene were changed to trigger object layers and the rocketbullfighterboss was modified to have a larger field of view